+++ src/state.ts
@@ -415,12 +415,12 @@ export class FormState<
   }
 
   @action
-  async processAll() {
+  async processAll(liveOnly?: boolean) {
     if (this.processor == null) {
       throw new Error("Cannot process all without backend configuration");
     }
 
-    return this.processor.realProcessAll();
+    return this.processor.realProcessAll(liveOnly);
   }
 
   @action